@@ -1,42 +1,40 @@
-﻿using AssParser.Lib;
+using AssParser.Lib;
 
-namespace AssParser
+namespace AssParser;
+
+public static class Program
 {
-    internal class Program
+    private static async Task Main(string[] args)
     {
-        static void Main(string[] args)
+        try
         {
-            try
+            var assFile = await Lib.AssParser.ParseFileAsync(ParserBenchmark.TestFileName);
+            var fonts = assFile.UsedFonts();
+            foreach (var font in fonts)
             {
-                var assfile = Lib.AssParser.ParseAssFile(@"[Nekomoe kissaten&VCB-Studio] Cider no You ni Kotoba ga Wakiagaru [Ma10p_1080p][x265_flac].jp&sc.ass").Result;
-                var fonts = assfile.UsedFonts();
-                foreach (var font in fonts)
-                {
-                    Console.WriteLine(font.FontName + "\t" + font.UsedChar);
-                }
-                var txt = assfile.ToString();
+                Console.WriteLine(font.FontName + "\t" + font.UsedChar);
             }
-            catch (AggregateException ae)
-            {
-                foreach (var ex in ae.InnerExceptions)
-                {
-                    // Handle the custom exception.
-                    if (ex is AssParserException)
-                    {
-                        var assExcption = ex as AssParserException;
-                        Console.WriteLine(assExcption?.ToString());
-                    }
-                    // Rethrow any other exception.
-                    else
-                    {
-                        Console.WriteLine(ex.Message);
-                    }
-                }
-            }
-#if !DEBUG
-            var summary = BenchmarkRunner.Run<ParserBenchmark>();
-#endif
-            Console.ReadLine();
+            var txt = await assFile.GetStringAsync();
         }
+        catch (AggregateException ae)
+        {
+            foreach (var ex in ae.InnerExceptions)
+            {
+                // Handle the custom exception.
+                if (ex is AssParserException assException)
+                {
+                    Console.WriteLine(assException?.ToString());
+                }
+                // Rethrow any other exception.
+                else
+                {
+                    Console.WriteLine(ex.Message);
+                }
+            }
+        }
+#if !DEBUG
+        var summary = BenchmarkDotNet.Running.BenchmarkRunner.Run<ParserBenchmark>();
+#endif
+        Console.ReadLine();
     }
 }
