@@ -1,1 +1,1 @@
-global using Xunit;
+global using Microsoft.VisualStudio.TestTools.UnitTesting;
